api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
cors==1.0.1
dnspython==2.5.0
